enemy damage implementation
implemented a way for enemies to deal damage to the player. the
structuring of the logic I am not super sure about but I cannot
currenlty think of a better way and I am not sure if I want to
implement it in the same way as the tutorial.

also added trees.