soft collision and map expansion
added in soft collision for the bats so that they will not sit on top of
each other when running.

added in a camera to use for following the character around the map.

working on expanding the map. added in some bounding cliffs and working
on adding paths. need to figure out how best to add in all the
foliage(?) and setup some enemies to harras the player.

need to figure out how best to deal with the UI since the actual assets
are so small that the ui elements are also super small and do not offer
a lot customization since they are near the minimum limit for some of
them. especially text.